First draft SVC page (#340)
A first draft Sustainability Vocabulary Catalog page. Including 

* A standard structure for scheme owners to publish their vocabularies
of conformity criteria. This makes sure that the criterion URI is
**correct.**
* A catalog of schemes that can be used as the entry point to discover
scheme-specific criteria. This makes sure that the criterion URI is
**discoverable.**
* A conformity topic taxonomy that can be used to classify criteria from
different schemes. This makes sure that different criterion URI are
**comparable.**

The description of the declarations structure that is used for claims in
DPP and DFR has moved to the DPP page.