Fix `mimeType` and `unitOfMeasure` external code table links so they concatenate with values (#383)
Update `mimeType` codetable link, removing `allTypes` as the external
examples follow the pattern `https://mimetype.io/application/atom+xml`
so the join of the mimeType is just with the root address.
Similarly, update unitOfMeasure codetable link, appending a `#` since
the external examples follow the pattern
`https://vocabulary.uncefact.org/UnitMeasureCode#10`

Note: Jargon uses the URL both for concatenating with a value for an ID
and for displaying to the user as the link where they can see all
values, but it can't work for both always. This change gives priority to
the former (valid and usable jsonld, with possibly incorrect link in
JSONSchema description to users). See
https://github.com/jargon-sh/issues/issues/52

Co-authored-by: ashleythedeveloper <60303491+ashleythedeveloper@users.noreply.github.com>