Sync the released 0.6.0 schemas, instances and vocabs. (#410)
With this change, there is no longer any references to `working` within
the data.

We currently only have `onSnapshot` syncs of data to the
data-models-staging branch, so when we released 0.6.0, although Jargon
then created the correct artefacts (and our existing release process
published those artefacts correctly), the data-models-staging branch is
not (currently) updated with the release changes.

So this PR just manually syncs the released artefacts to the
data-models-staging branch before we merge that branch into main and
re-sync it to HEAD to start the next snapshots for the next release.

```console
(release-0.6.0-fix) ~/dev/gosource/spec-untp/data-models$ rg working
(release-0.6.0-fix) ~/dev/gosource/spec-untp/data-models$ 
```