Update IdentityResolver.md
I provided the creation of a product did, the DPP deep link URL, what is in a QR code and how the DID is resolved. How it works with DIDs is scattered all over the place so I think it would make sense to restructure the page. Then how is works with dids could be in one section. Also the coding examples should be updated and included in the same section. Coding examples can be found here:  https://medium.com/spherity/the-worlds-first-decentralized-verifiable-battery-passport-a0f42c1bcb5e

Signed-off-by: susanne-guth-orlowski <just@4therecord.io>